chore: Add separate page for raw elements
